introduce broker manager and shutdown hook for test container
