+++ module/communication/mqtt/build.gradle.kts
@@ -8,6 +8,7 @@ dependencies {
     api(project(":jul.extension.type.processing"))
     api("com.hivemq:hivemq-mqtt-client:_")
     implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:_")
+    testImplementation(project(":jul.communication.mqtt.test"))
     testImplementation("org.testcontainers:junit-jupiter:_") {
         exclude(group = "junit", module = "junit")
     }
